forras

+++ Project/admin/edit_user.php
@@ -1,29 +1,31 @@
 <?php include("includes/header.php");
 
-if(!$session->isSignedIn()){
+if (!$session->isSignedIn()) {
     redirect("login.php");
 }
 
 $message = "";
-if(empty($_GET['id'])){
+if (empty($_GET['id'])) {
     redirect('users.php');
-}else{
+} else {
     $user = user::findById($_GET['id']);
-    $oldFile = SITEROUTE . DS . 'admin' .DS . $user->Path();
-    if(isset($_POST['update'])){
-        if($user){
+    $oldFile = SITEROUTE . DS . 'admin' . DS . $user->Path();
+    if (isset($_POST['update'])) {
+        if ($user) {
             $user->first_name = $_POST['first'];
             $user->last_name = $_POST['last'];
             $user->username = $_POST['username'];
             $user->password = $_POST['pass'];
-            if(empty($_FILES['file']['name'])){
+            if (empty($_FILES['file']['name'])) {
                 $user->save();
-            }else{
+            } else {
                 $user->setFiles($_FILES['file']);
-                if($user->saveFileToo()){
-                    unlink($oldFile);
+                if ($user->saveFileToo()) {
+                    if (!strpos($oldFile, 'http://placehold.it/100X100&text=image')) {
+                        unlink($oldFile);
+                    }
                     $message = "Photo uploaded successfully";
-                } else{
+                } else {
                     $message = join("<br>", $user->customErrors);
                 }
             }
@@ -46,7 +48,7 @@ if(empty($_GET['id'])){
                     <small>Subheading</small>
                 </h1>
                 <div class="col-md-2">
-                    <img class="img-responsive photoSize" src="<?= $user->path()?>" alt="">
+                    <img class="img-responsive photoSize" src="<?= $user->path() ?>" alt="">
                 </div>
                 <form action="" method="post" enctype="multipart/form-data">
                     <div class="col-md-6">
@@ -57,28 +59,28 @@ if(empty($_GET['id'])){
 
                         <div class="form-group">
                             <label for="username">Username</label>
-                            <input type="text" name="username" class="form-control" value="<?= $user->username?>" >
+                            <input type="text" name="username" class="form-control" value="<?= $user->username ?>">
                         </div>
 
                         <div class="form-group">
                             <label for="first">First name</label>
-                            <input type="text" name="first" class="form-control" value="<?= $user->first_name?>">
+                            <input type="text" name="first" class="form-control" value="<?= $user->first_name ?>">
                         </div>
                         <div class="form-group">
                             <label for="last">Last name</label>
-                            <input type="text" name="last" class="form-control" value="<?= $user->last_name?>">
+                            <input type="text" name="last" class="form-control" value="<?= $user->last_name ?>">
                         </div>
                         <div class="form-group">
                             <label for="pass">Password</label>
-                            <input type="password" name="pass" class="form-control" value="<?= $user->password?>">
+                            <input type="password" name="pass" class="form-control" value="<?= $user->password ?>">
                         </div>
                         <div class="form-group">
                             <input type="submit" name="update" value="Update" class="btn btn-primary pull-right">
                         </div>
                         <div class="form-group">
-                            <a  href="users.php?id=<?php echo $user->id;?>&path=users.php&class=user" class="btn btn-danger">Delete</a>
+                            <a href="users.php?id=<?php echo $user->id; ?>&path=users.php&class=user"
+                               class="btn btn-danger">Delete</a>
                         </div>
-
                     </div>
                 </form>
             </div>
